Version 2 - Much Improved Saving, Naming, and Everything Else!
Added version 2 of the script, which saves images in separate directories according to subreddit, names images according to post title, and is friendly to running the script multiple times (i.e. doesn't redownload images).

The script is now configurable via settings.txt.

I also added encodings to the script files, which should help Mac compatibility.

With V2 actually downloading submissions, I can expand the features of the script to include downloading saved comments, creating a list of submissions which couldn't be saved, and other cool stuff I haven't thought of. This features are TBD, but now much easier to implement in the future.