Added Support for RedditUploads
All links which have a supported Content-Type will now be downloaded (instead of always relying on the file type being in the URL). This should mean all reddituploads links will work.

Images with filetypes in the URL that do not match the content type will be corrected. This should reduce the number of files saved under incorrect formats.

I also added a simple statistics output so users know how many images were saved/unsupported/already saved etc.

Imgur albums are the next big ordeal. I'm going to have to add the imgur API to get those, but it should be nice once it's in.