Added Early Out for already saved submissions
I added a new feature which compares the newest submissions which were last saved so that the script can early-out once it reaches them. This saves a significant amount of API requests and download time. 

The feature is enabled by default, but can be disabled via setting Reddit_Try_Request_Only_New and/or Tumblr_Try_Request_Only_New to False in settings.txt