Added Gfycat API support, changed certificates
* Certificates are now more appropriately named (liked_saved_server
instead of server_jupyter_based). The selection code is backwards compatible
* Added dependency on py-gfycat. The version in Pip is very
out-of-date, so I instead install it straight from Github. I added
Gfycat settings and instructions. Gfycat URLs are parsed for their
gfynames, then the MP4 URL is retrieved from Gfycat through the API
* #22: Auto-create output_dir if it doesn't already exist. Image Saver
was doing this but the parent script was not