Added feature to retry downloading content
Because I recently added video downloading support, there should be
a lot of content which the script can now download which it previously
couldn't.

I made it so that the "Content which failed to download"
page now has a button to let you retry to download content. This is a
pretty sweet change, though the interface isn't amazing yet.

* Already downloaded videos are removed from unsupported list (#54)
* Successful submission downloads are removed from Unsupported
  list (close #54)
* Convert imgur indirect links via the imgur API, instead of hacking
the HTML. This is especially necessary now that login is required for
NSFW imgur images
* Pipes for subprocesses now handle exceptions, giving the user
helpful output if the server stumbles (related to #37)
* Fixed nav elements
* Remove submissions which have files associated with them on legacy
unsupported submissions import, to close #63