isRedditGallery needs reddit client to determine whether submission is gallery:
sometimes submissions may have a bodyUrl which is the permalink to a
post, but which point to a reddit gallery (and have content type HTML).
if this is the case, the only way to determine whether a post is a
gallery is not with bodyUrl or postUrl but by asking `praw.submission`
for the `post.url` - which will contain `/gallery/` is this is a
gallery.

also:
- move urlContentType to before isRedditGallery()
- rename `client` variable to `reddit_client`